move genre and beat select to the app component
The sequencer component was doing two things : rendering the sequencer and
managing the beat library. The two app-select-input controls, along with
genreChange, beatChange and selectBeat, now live in the app component.

No input/output wiring is needed : SequencerService is providedIn root, so
both components already read the same genre and beat state. The initial
library load (initialize + first genre selection) moves up as well, since it
belongs to library management rather than to the sequencer.

The selects are declared once in an ng-template and projected into both the
desktop and the mobile landscape branch.
