+++ frontend/src/app/ui/components/sequencer/sequencer.component.scss
@@ -47,10 +47,6 @@ $gap: 4px;
     }
   }
 
-  div+div {
-    margin-top: 8px;
-  }
-
   .bpm-input-group {
     align-items: center;
     gap: 16px;
